rearrange fingerprints in class hierarchy
